feat: 임시구현한 API에 web3j 연결 추가

+++ rabbit-server/src/main/java/com/rabbit/loan/domain/dto/response/LentSummaryResponseDTO.java
@@ -15,7 +15,7 @@ import java.math.BigInteger;
 @AllArgsConstructor
 public class LentSummaryResponseDTO {
 
-    private int totalIncomingLa;
-    private int monthlyIncomingLa;
+    private Long totalIncomingLa;
+    private Long monthlyIncomingLa;
     private String nextIncomingDt;
 }
